refactored raspi wolfboot_printf proposal
